refactor(ofm_core): peel transfer and loan execution out of transfers
Second peel. Moves the seven functions that carry out a deal once its terms
are agreed — execute_transfer, execute_loan, the two pending-registration
reservations, the buyer affordability check, fee rounding and feedback
building — into transfers/execution.rs.

The bodies are byte-identical to what was there; only the visibility
keyword changed, since the negotiation code in mod.rs still calls them. The
submodule takes the parent's scope with `use super::*`, so no import block
had to be reconstructed.

transfers/mod.rs 3671 -> 3313 lines.
